requirements.txt
bpy==3.6.0
einops==0.7.0
ffmpeg-python==0.2.0
flash-attn==2.6.3
gymnasium==0.29.1
h5py==3.10.0
pillow==10.2.0
matplotlib==3.7.4
numpy==1.24.4
open3d==0.18.0
opencv-python==4.9.0.80
pandas==2.0.3
randomname==0.2.1
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
timm==0.9.16
tqdm==4.66.1
transformers[torch]==4.45.2
transforms3d==0.4.1
yt-dlp==2023.12.30
wandb==0.16.6